Add macro
