+++ src/babel.js
@@ -13,9 +13,8 @@ import {
   addClassName,
   getScope,
   processCss,
-  combinePlugins,
-  booleanOption,
-  createReactComponentImportDeclaration
+  createReactComponentImportDeclaration,
+  setStateOptions
 } from './_utils'
 
 import { STYLE_COMPONENT } from './_constants'
@@ -265,24 +264,7 @@ export default function({ types: t }) {
           state.ignoreClosing = null
           state.file.hasJSXStyle = false
 
-          const vendorPrefixes = booleanOption([
-            state.opts.vendorPrefixes,
-            state.file.opts.vendorPrefixes
-          ])
-          state.opts.vendorPrefixes =
-            typeof vendorPrefixes === 'boolean' ? vendorPrefixes : true
-          const sourceMaps = booleanOption([
-            state.opts.sourceMaps,
-            state.file.opts.sourceMaps
-          ])
-          state.opts.sourceMaps = Boolean(sourceMaps)
-
-          if (!state.plugins) {
-            state.plugins = combinePlugins(state.opts.plugins, {
-              sourceMaps: state.opts.sourceMaps,
-              vendorPrefixes: state.opts.vendorPrefixes
-            })
-          }
+          setStateOptions(state)
         },
         exit({ node, scope }, state) {
           if (
